`tools/generator-go-sdk`: moving `featureflags` into the `internal` folder
