+++ tools/generator-go-sdk/generator/templater_models.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/hashicorp/pandora/tools/generator-go-sdk/featureflags"
+	"github.com/hashicorp/pandora/tools/generator-go-sdk/internal/featureflags"
 	"github.com/hashicorp/pandora/tools/sdk/resourcemanager"
 )
