+++ tools/generator-go-sdk/internal/featureflags/flags.go
